refactor/be/#157 admin board repository 추가
