+++ admin/src/main/java/kernel/maidlab/admin/board/repository/AdminBoardRepository.java
@@ -1,0 +1,9 @@
+package kernel.maidlab.admin.board.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import kernel.maidlab.common.entity.board.Board;
+
+public interface AdminBoardRepository extends JpaRepository<Board, Long> {
+
+}
