+++ admin/src/main/java/kernel/maidlab/admin/manager/controller/AdminManagerApi.java
@@ -2,6 +2,7 @@ package kernel.maidlab.admin.manager;
 
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
